requirements.txt
numpy>=1.21
pandas>=1.3
pyyaml>=6.0
tqdm>=4.60
